add test and example
